feat: organize CLI tests so that they can be invoked programatically
