+++ cmd/crates/soroban-test/tests/it/main.rs
@@ -1,9 +1,10 @@
 mod arg_parsing;
 mod config;
 
-mod contract_sandbox;
 mod custom_types;
 mod dotenv;
+mod hello_world;
+mod integration_and_sandbox;
 mod lab;
 mod plugin;
 mod util;
